Move logic from module class to plugin
