+++ spring-xd-module/src/main/java/org/springframework/xd/module/spark/streaming/NonBindingResourceConfiguredModule.java
@@ -1,0 +1,41 @@
+/*
+ * Copyright 2015 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.springframework.xd.module.spark.streaming;
+
+import org.springframework.xd.module.ModuleDeploymentProperties;
+import org.springframework.xd.module.ModuleDescriptor;
+import org.springframework.xd.module.core.ResourceConfiguredModule;
+import org.springframework.xd.module.options.ModuleOptions;
+
+/**
+ * A module type that chooses not to bind to the MessageBus. Used in the Spark case for example.
+ *
+ * @author Eric Bottard
+ *
+ */
+public class NonBindingResourceConfiguredModule extends ResourceConfiguredModule {
+
+	public NonBindingResourceConfiguredModule(ModuleDescriptor descriptor, ModuleDeploymentProperties deploymentProperties,
+											  ClassLoader classLoader, ModuleOptions moduleOptions) {
+		super(descriptor, deploymentProperties, classLoader, moduleOptions);
+	}
+
+	@Override
+	public boolean shouldBind() {
+		return false;
+	}
+}
